pyproject.toml
[tool.poetry]
name = "kittbot"
version = "0.1.0"
description = "Greetings. I am KITT, the Knight Industries Two Thousand. I am a highly advanced, artificially intelligent, self-aware vehicle. I am here to manage the Discord server for Improv Knight, a VRChat-based improv group."
authors = ["polyjitter <polyjitter@fastmail.com>"]
license = "GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
hikari = "^2.0.0.dev119"
hikari-toolbox = "^0.1.5"
hikari-miru = "^3.1.1"
toml = "^0.10.2"
hikari-lightbulb = "^2.3.3"
uvloop = [
    {version = "^0.17.0", platform = "linux"},
    {version = "^0.17.0", platform = "darwin"},
]
sqlalchemy = {extras = ["asyncio"], version = "^2.0.15"}
aiosqlite = "^0.19.0"
datefinder = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
